Update game logic - Part I
I've added the logic to swap from the choices gameplay to the console output gameplay.

There's a bug where the output is still outputting even though the prompt description is disabled from the console window and the button to proceed pops up 30 secs or so afterwards. Next update should focus on getting back from the prompt and updating the rooms available.